Added input validation and error handling
The EncryptionStrategy interface now has a method called getError().
This returns a String if there was an error and returns null if
everything went fine. I added the error string and implemented it in the
strategies.

Also commented the controller, view, shift, and substitution classes.
The view also now asks the user to re-enter their selections if they're
invalid.